Changed Tab Size to 2 and removed extra newlines
